refactor: extract checklist evaluators into their own module
Move the pure evaluation logic (state types, thresholds, stablecoin
set, and the six evaluate*() functions) out of PositionDetails.tsx
into src/components/dashboard/monitoringChecklist.ts. PositionDetails
now only owns the React rendering and badge-variant mapping; the
evaluators are decoupled from the component and reusable/testable on
their own.

Also adds evaluateHealthFactor / evaluateLtv so all six checklist
items go through the same shape, replacing the inline objects in
MonitoringChecklistCard.
